pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alpaca-options-quant"
version = "0.1.0"
description = "No-news quant options paper-trader with Discord alerts"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2",
  "pandas>=2",
  "numpy>=1.26",
  "PyYAML>=6",
  "SQLAlchemy>=2.0",
  "tenacity>=8",
  "pytz>=2024.1",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7",
  "types-python-dateutil>=2",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.package-data]
"src" = ["**/*.yaml"]

